Update Grade 7 Histogram
Since g1<0 there is some left skewedness.  However since |g1|<3.41 the data is not substantially skewed.
Since g2>0 the data is leptokurtic.  Since |g2|<6.81 the data is not substantially leptokurtic.
I am surprised that the data is not substantally leptokurtic because there appears to be a definitive peak and substantial drops to both sides of the peak.  The peak is lower than those found in the grades 5 and 6 histograms and while the bars at the high end are smaller the amount of level 1 performances decreased.